Move resolveReference logic to classes.ts
