+++ src/classes.ts
@@ -1,13 +1,13 @@
 import { getKey, readJsonFile } from './data';
 import {
     capitalize,
+    checkForDisallowedSymbols,
     cleanDNDText,
     Description,
     DescriptionType,
     parseAbilityScore,
     parseClassResourceValue,
     parseDescriptions,
-    resolveReferences,
     title,
 } from './parser';
 import { getClassesUrl } from './urls';
@@ -579,6 +579,167 @@ class CharacterClass {
         this.subclassUnlockLevel = lowestLevel === 999 ? null : lowestLevel;
         this.subclassLevelFeatures = result;
     }
+}
+
+function resolveClassFeatReference(
+    text: string,
+    feats: ClassFeatureDictionary | null,
+    type: 'refClassFeature' | 'refSubclassFeature'
+): Description[] {
+    const regex = new RegExp(`\\{#${type}\\s+([^}]+)\\}`, 'g');
+    const matches = [...text.matchAll(regex)];
+
+    const onlyOneReference = matches.length === 1;
+    const trimmedText = text.trim();
+    const trimmedMatch = onlyOneReference ? matches[0][0].trim() : '';
+    const isPureReference = onlyOneReference && trimmedText === trimmedMatch;
+
+    const isOnlyReferences =
+        matches.length > 0 &&
+        text
+            .replace(regex, '')
+            .trim()
+            .replace(/^[•\s\r\n]+|[•\s\r\n]+$/g, '') === '';
+
+    // Case 1: String with only a {#ref } and nothing else.
+    if (isPureReference || isOnlyReferences) {
+        const parts = matches[0][1].split('|').map((p) => p.trim());
+
+        let name: string, className: string, source: string, levelStr: string;
+        let subclassName: string | undefined, subclassSource: string | undefined;
+
+        if (type === 'refClassFeature') {
+            [name, className, source, levelStr] = parts;
+        } else {
+            [name, className, source, subclassName, subclassSource, levelStr] = parts;
+        }
+
+        const featSource = source || 'PHB';
+        const key = getKey(className, featSource);
+
+        if (!feats || !feats[key]) {
+            return [
+                {
+                    name: '',
+                    type: DescriptionType.text,
+                    value: `${BulletPoint} ${getKey(name, featSource)}`,
+                },
+            ];
+        }
+
+        const feat = feats[key].find(
+            (f) => f.name.trim().toLowerCase() === name.trim().toLowerCase()
+        );
+
+        if (!feat?.descriptions) throw `Could not find ${type} for ${key}`;
+        let descs = feat.descriptions.map((d) => ({ ...d }));
+        descs[0].value = `__${getKey(name, featSource)}__: ${descs[0].value}`;
+
+        return descs;
+    }
+
+    // Case 2: Inline substitution (multiple references or formatting like bullets)
+    const updatedText = text.replace(regex, (match, inner) => {
+        const parts = inner.split('|').map((p: string) => p.trim());
+
+        const [name, , source] = parts;
+        const featSource = source || 'PHB';
+        return getKey(name, featSource);
+    });
+
+    return [
+        {
+            name: '',
+            type: DescriptionType.text,
+            value: updatedText,
+        },
+    ];
+}
+
+function resolveOptionalFeatReference(entry: Description): Description {
+    const value = entry.value as string;
+
+    const updatedValue = value.replace(
+        /\{#refOptionalfeature\s+([^|}]+)(?:\|([^}]+))?\}/g,
+        (_, name: string, source?: string) => {
+            const finalSource = source?.trim() || 'PHB';
+            return `${name.trim()} (${finalSource})`;
+        }
+    );
+
+    return {
+        ...entry,
+        value: updatedValue,
+    };
+}
+
+function resolveReferences(
+    entries: Description[],
+    classFeats: ClassFeatureDictionary | null = null,
+    subclassFeats: ClassFeatureDictionary | null = null,
+    isSubResolve: boolean = false
+): Description[] {
+    const resolvedEntries: Description[] = [];
+
+    for (const entry of entries) {
+        if (entry.type === DescriptionType.text) {
+            const text = entry.value as string;
+
+            if (!text.includes(`{#`)) {
+                resolvedEntries.push(entry);
+                continue;
+            }
+
+            if (text.includes(`refClassFeature`))
+                resolvedEntries.push(
+                    ...resolveClassFeatReference(text, classFeats, 'refClassFeature')
+                );
+            else if (text.includes(`refSubclassFeature`))
+                resolvedEntries.push(
+                    ...resolveClassFeatReference(text, subclassFeats, 'refSubclassFeature')
+                );
+            else if (text.includes('refOptionalfeature'))
+                resolvedEntries.push(resolveOptionalFeatReference(entry));
+            else throw `Unsupported text-description reference-type in: ${text}`;
+        } else if (entry.type === DescriptionType.table) {
+            resolvedEntries.push(entry); // For now, tables don't have any references.
+        } else {
+            throw `Could not resolve unsupported DescriptionType ${entry.type}`;
+        }
+    }
+
+    if (!isSubResolve) {
+        let indexesToResolve: number[] = [];
+        let entriesToSubResolve: Description[] = [];
+        for (let i = 0; i < resolvedEntries.length; i++) {
+            const e = resolvedEntries[i];
+            if (e.type === DescriptionType.text) {
+                const val = e.value as string;
+                if (!val.includes('{#')) continue;
+
+                indexesToResolve.push(i);
+                entriesToSubResolve.push(e);
+            }
+        }
+
+        for (let j = indexesToResolve.length - 1; j >= 0; j--) {
+            const index = indexesToResolve[j];
+            const resolved = resolveReferences(
+                [entriesToSubResolve[j]],
+                classFeats,
+                subclassFeats,
+                true
+            );
+            resolvedEntries.splice(index, 1, ...resolved);
+        }
+    }
+
+    for (const resolvedEntry of resolvedEntries) {
+        if (resolvedEntry.type !== DescriptionType.text) continue;
+        checkForDisallowedSymbols(resolvedEntry.value as string);
+    }
+
+    return resolvedEntries;
 }
 
 export function getClasses(): any[] {
